requirements.txt
Flask[async]==3.0.3
flask_socketio==5.3.6
flask_session==0.8.0
flask_cors==4.0.1
flask_mailman==1.1.1

openai==1.35.4
python-dotenv==1.0.1
mongoengine==0.28.2
Authlib==1.3.1
redis==5.0.6
requests==2.32.3
json_repair==0.25.2

torch~=2.3.1
numpy~=1.26.4
transformers~=4.42.3
scikit-learn~=1.5.1